requirements.txt
aiogram>=3.13.0
aiohttp>=3.9.0
brotli>=1.1.0
certifi>=2024.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
google-generativeai>=0.8.0
lxml>=5.0.0
beautifulsoup4>=4.12.0
transliterate>=1.10.0
libgen-api>=1.0.0
